Correção no envio da fila e na mensagem do erro de validação
