+++ src/Domain/FundTransfer.Domain/Repositories/ITransferRepository.cs
@@ -1,10 +1,10 @@
-﻿using FundTransfer.Domain.Dtos;
+﻿using FundTransfer.Domain.Entities;
 
 namespace FundTransfer.Domain.Repositories
 {
     public interface ITransferRepository
     {
-        Task AddAsync(Transfer transferDto, CancellationToken cancellationToken);
+        Task AddAsync(Transfer transfer, CancellationToken cancellationToken);
         Task<Transfer> GetAsync(Guid transactionId, CancellationToken cancellationToken);
     }
 }
